requirements.txt
aiohttp
aiodns
aiohttp-retry
tenacity
fastapi
google-adk
google-auth
google-cloud-aiplatform
google-cloud-storage
python-dotenv
uvicorn
json-repair

# Authentication dependencies
python-jose[cryptography]
passlib[bcrypt]
python-multipart
email-validator

# MongoDB dependencies
motor
pymongo
dnspython
certifi
